server/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==23.0.0
load-dotenv==0.1.0
packaging==24.2
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
psycopg2== 2.9.10
django-cloudinary-storage==0.3.0
cloudinary==1.42.2